requirements.txt
ultralytics>=8.0.0
torch>=1.7.0
opencv-python>=4.1.2
pyyaml>=5.3.1
numpy>=1.19.0
scikit-learn>=0.24.0